# internal/adapters/http/dto/user.go
package dto

type User struct {
	ID          int      `json:"id" validate:"gte=0"`
	FirstName   string   `json:"first_name"`
	LastName    string   `json:"last_name"`
	Email       string   `json:"email"`
	Password    string   `json:"password"`
	Preferences []string `json:"preferences"`
}

type RegisteredUser struct {
	FirstName   string   `json:"first_name" validate:"required"`
	LastName    string   `json:"last_name" validate:"required"`
	Email       string   `json:"email" validate:"required,email"`
	Password    string   `json:"password" validate:"required,min=8,max=64,passwod_strength"`
	Preferences []string `json:"preferences" validate:"required,unique,user_preferences_check"`
}

type RegisterVerify struct {
	SentOtpbyUser string `json:"otp" validate:"required"`
}

type LoginUser struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}
type DeleteUser struct {
	UserID int `json:"user_id" validate:"required,gte=0"`
}